requirements.txt
langchain>=0.1.0
langchain-openai>=0.0.2
langchain-community>=0.0.28
openai>=1.3.0
fastapi>=0.104.0
uvicorn>=0.24.0
python-multipart>=0.0.6
pydantic>=2.4.2
chromadb>=0.4.18
python-dotenv>=1.0.0
tiktoken>=0.5.1
pandas>=2.1.1
numpy>=1.24.3
pillow>=10.0.0
python-jose>=3.3.0
passlib>=1.7.4
bcrypt>=4.0.1
websockets>=12.0
aiofiles>=23.2.1
jinja2>=3.1.3
httpx>=0.27.0
pytest>=8.0.2
pytest-asyncio>=0.23.5
pytest-cov>=4.1.0
black>=24.2.0
isort>=5.13.2
mypy>=1.8.0
ruff>=0.2.2 